drakrun: support network in installation process
